refactor breaches util functions
